requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
decorator==4.0.11
Django==1.10.6
django-ckeditor==5.3.0
django-extensions==1.7.7
django-filter==1.0.2
django-js-asset==0.1.1
django-rest-swagger==2.1.2
djangorestframework==3.6.2
djoser==0.6.0
idna==2.5
ipython==5.3.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.9.6
kombu==4.0.2
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.13
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2016.10
requests==2.18.3
rules==1.2.1
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
traitlets==4.3.2
Unipath==1.1
uritemplate==3.0.0
urllib3==1.22
vine==1.1.3
wcwidth==0.1.7
